refactor: move AppBarBackButton component files to their own folder
